Added capture button and gallery thumbnail widgets,
